Correct identification of destructor rules
Destruction rules are those that take an injective fact in their
premises and do not produce the fact with the same ith term argument
in their conclusions. Previously, the implementation would have
ignored destructor rules which manipulated another instance of the
injective fact with a different ith term argument.

This code filters each rule down to their injective fact candidates,
extracts their candidate ith term argument and checks whether it
appears in both the premises and conclusions -- if any of the
premises are not in the conclusions, then the rule is a destructor.

This code allows the rule [ St(~a), Fr(~b) ] --> [ St(~b) ] to be
correctly identified as both a constructor and destructor rule
(assuming St is an injective fact).