Reduce restrictions on injective fact rule conclusions.
Previously, Tamarin has only allowed single copies of injective
facts to appear in conclusions. However, this restriction can be
reduced to single copies of the fact and the fresh term.

This code does that by detecting whether the injective facts fresh
terms are duplicated.

This allows rules such as [ St(~a), St(~b) ] -> [ St(~a), St(~b) ]
to be written without preventing St from being detected as
injective.