Add a representation of invariant facts as a type
This adds a new type of fact - InvFact, which stores the invariant
values found in protocol facts. All invariant facts are linear,
with arity > 1 and are currently prefixed with "Inv" in the UI.

Includes a constructor and a method to convert existing protoFacts
into an invariant fact, with an optional list of indices for
invariant term positions.